requirements.txt
translatekit
python-dotenv
